requirements.txt
Flask==3.0.0
Werkzeug==3.1.3
psycopg2-binary==2.9.9
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
gunicorn==21.2.0
python-dotenv==1.0.0